shared & favourites

+++ packages/chat/src/components/ChatSessionList.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-empty-function */
-import React, { ReactNode, useCallback, useEffect, useMemo, useState } from "react";
+import React, { ReactNode, SyntheticEvent, useCallback, useEffect, useMemo, useState } from "react";
 import { FixedSizeList, ListChildComponentProps } from "react-window";
 import InfiniteLoader from "react-window-infinite-loader";
 import { useChatHistory, useChatModels, useChatSession, useChatSessionFavourite } from "../api/chat";
@@ -28,11 +28,12 @@ export function ChatSessions(props: ChatSessionListProps) {
   const queryClient = useQueryClient();
   const axios = useAxiosPrivate();
   // Array of items loaded so far.
-  const [items, setItems] = useState<IChatSession[]>([]);
-  const fetchChatSessions = async ({ pageParam = 0 }) => {
-    const res = await axios('/api/chat/history?limit=10&offset=' + (pageParam*10));
-    return res.data.data;
-  };
+  const [items, setItems] = useState<{id:string}[]>([]);
+  const fetchChatSessions = useMemo(()=>(async ({ pageParam = 0 }) => {
+      const res = await axios(`/api/chat/history?limit=10&offset=${pageParam*10}&type=${props.type}`);
+      return res.data.data;
+    }
+  ),[axios, props.type]);
   const {
     data,
     error,
@@ -42,7 +43,7 @@ export function ChatSessions(props: ChatSessionListProps) {
     isFetchingNextPage,
     status,
   } = useInfiniteQuery({
-    queryKey: ['chathistory'],
+    queryKey: ['chathistory', props.type],
     queryFn: fetchChatSessions,
     getNextPageParam: (lastPage, pages) =>{
       if (lastPage.length < 5) return undefined;
@@ -51,8 +52,8 @@ export function ChatSessions(props: ChatSessionListProps) {
   })
 
   useEffect(()=>{
-    const newItems: IChatSession[] = [];
-    data?.pages.forEach((group)=>{ group.forEach((item:IChatSession)=>{newItems.push(item)})});
+    const newItems: {id:string}[] = [];
+    data?.pages.forEach((group)=>{ group.forEach((item:{id:string})=>{newItems.push(item)})});
     setItems(newItems);
   },[data])
 
@@ -66,62 +67,6 @@ export function ChatSessions(props: ChatSessionListProps) {
 
   // Every row is loaded except for our loading indicator row.
   const isItemLoaded = useCallback((index: number) => (!hasNextPage || index < items.length),[hasNextPage, items.length]);
-  // Render an item or a loading indicator.
-  const Item:any = (props:{ index:number, style:any }) => {
-    if (!isItemLoaded(props.index)) {
-      return <div style={props.style}>{""}</div>;
-    }
-    const session = items[props.index];
-
-    return <Box key={session.id+'i'}
-      sx={{mb:1, p:0.5, minHeight:56,display:'flex', flexDirection:'row', alignItems:'center', cursor: 'pointer',
-        borderRadius:'3px',border:'1px solid #ccc', '&:hover':{borderColor:'#000'}, height:'78px',
-      }}>
-        {/* <IconButton size="small">{props.icon}</IconButton> */}
-        <Box sx={{flexGrow:1, display:'flex', flexDirection:'column', alignItems: 'stretch', pl:0.5}} >
-          <Typography variant="body2">{session.name}</Typography>
-          <Stack direction="row" display={'flex'} justifyContent="end" alignItems={'flex-end'}  spacing={1} sx={{width:'100%'}}>
-            <Typography variant="caption" sx={{flexGrow:1, color:'text.secondary' }}>{formatDistanceToNow(parseJSON(session.updatedAt),{addSuffix:true})}</Typography>
-            <IconButton aria-label="toggle favourite" size="small">
-            {session.type==='f'?<FavoriteIcon fontSize="inherit" color="secondary" />:<FavoriteBorderIcon fontSize="inherit" />}
-            </IconButton>
-            <IconButton aria-label="toggle sharing" size="small">
-            {session.type==='public'?<ShareIcon fontSize="inherit" color="success" />:<ShareIcon fontSize="inherit" />}
-            </IconButton>
-            <IconButton aria-label="delete session" size="small">
-              <DeleteIcon fontSize="inherit" />
-            </IconButton>
-          </Stack>
-        </Box>
-    </Box>
-  };
-  function renderRow(props: ListChildComponentProps) {
-    const { index, style } = props;
-    const session = items[props.index];
-
-    if (!session) return <p>Loading...</p>
-
-    return (
-      <ListItem style={style} key={index} component="div" disablePadding>
-        <ListItemButton>
-          <ListItemText primary={session.name}
-            secondary={<Stack direction="row" display={'flex'} justifyContent="end" alignItems={'center'}  spacing={1} sx={{width:'100%'}}>
-            <Typography variant="caption" sx={{flexGrow:1, color:'text.secondary' }}>{formatDistanceToNow(parseJSON(session.updatedAt),{addSuffix:true})}</Typography>
-            <IconButton aria-label="toggle favourite" size="small">
-            {session.type==='f'?<FavoriteIcon fontSize="inherit" color="secondary" />:<FavoriteBorderIcon fontSize="inherit" />}
-            </IconButton>
-            <IconButton aria-label="toggle sharing" size="small">
-            {session.type==='public'?<ShareIcon fontSize="inherit" color="success" />:<ShareIcon fontSize="inherit" />}
-            </IconButton>
-            <IconButton aria-label="delete session" size="small">
-              <DeleteIcon fontSize="inherit" />
-            </IconButton>
-          </Stack>}
-          />
-        </ListItemButton>
-      </ListItem>
-    );
-  }
 
   return status === 'loading' ? (
     <p>Loading...</p>
@@ -129,20 +74,9 @@ export function ChatSessions(props: ChatSessionListProps) {
     <Alert severity="error">Unable to load the chat sessions.</Alert>
   ) : (
     <>
-      {/* <FixedSizeList
-        height={400}
-        width={260}
-        itemSize={78}
-        itemCount={itemCount}
-        overscanCount={5}
-      >
-        {renderRow}
-      </FixedSizeList> */}
-      <React.Fragment>
-        {items.map((session:IChatSession,i: number) => (
-          <SessionSummary key={session.id} sessionid={session.id} onDelete={()=>queryClient.invalidateQueries({ queryKey: ['chathistory'] })}/>
-        ))}
-      </React.Fragment>
+      {items.map((session:{id:string},i: number) => (
+        <SessionSummary key={session.id} sessionid={session.id} onDelete={()=>queryClient.invalidateQueries({ queryKey: ['chathistory'] })}/>
+      ))}
       <div>
         <Button
           onClick={() => fetchNextPage()}
@@ -167,16 +101,19 @@ export function SessionSummary(props: {sessionid?: string, session?:IChatSession
   const {data:isFavourite, mutation:favMutation, invalidateCache: favInvalidateCache} = useChatSessionFavourite(session?.id);
   const navigate = useNavigate();
 
-  const deleteSession = useCallback(() => {
+  const deleteSession = useCallback((e: SyntheticEvent) => {
+    e.stopPropagation();
+    if (!session || isDeleted) return;
     axios.delete(`/api/chat/${session?.id}`)
     .then(() => {
       setIsDeleted(true);
       // if (props.onDelete) props.onDelete();
     })
     .catch((err)=>console.error(err))
-  },[axios, session?.id]);
+  },[axios, isDeleted, session]);
 
-  const toggleFavourite = useCallback(() => {
+  const toggleFavourite = useCallback((e: SyntheticEvent) => {
+    e.stopPropagation();
     if (!session) return;
     favMutation.mutateAsync(!isFavourite)
       .then(() => {
@@ -185,17 +122,21 @@ export function SessionSummary(props: {sessionid?: string, session?:IChatSession
       .catch((err) => {
         console.error(err);
       });
+      return true;
   },[favInvalidateCache, favMutation, isFavourite, session]);
 
-  const toggleSharing = useCallback((e:any) => {
-    e.preventDefault();
-    axios.post(`/api/chat/${session?.id}/sharing`,{
-      type: session?.type==='public'?'private':'public'
+  const toggleSharing = useCallback((e: SyntheticEvent) => {
+    e.stopPropagation();
+    if (!session) return;
+    axios.post(`/api/chat/${session.id}/sharing`,{
+      type: session.type==='public'?'private':'public'
     })
     .then(() => {
       invalidateCache();
     })
-    .catch((err)=>console.error(err))
+    .catch((err)=>console.error(err));
+
+    return true;
     // mutation.mutateAsync({
     //   id: session?.id,
     //   type: session?.type==='public'?'private':'public'
@@ -211,7 +152,9 @@ export function SessionSummary(props: {sessionid?: string, session?:IChatSession
   return <Fade in timeout={500}><Box sx={theme=>({
       mb: 1, p: 0.5, minHeight: 56, display: 'flex', flexDirection: 'row', alignItems: 'center', cursor: 'pointer',
       borderRadius: '3px', border: `1px solid #ccc`, borderColor:theme.palette.divider, '&:hover': { borderColor: theme.palette.action.active },
-    })} onClick={()=>!isDeleted && navigate(`/chat/${session.id}`)}>
+    })} onClick={()=>{
+      !isDeleted && navigate(`/chat/${session.id}`)
+    }}>
     {/* <IconButton size="small">{props.icon}</IconButton> */}
     <Box sx={{ flexGrow: 1, display: 'flex', flexDirection: 'column', alignItems: 'stretch', pl: 0.5 }}>
       <Typography variant="body2" sx={isDeleted?{textDecoration: "line-through"}:{}} >{session.name}</Typography>
